# Reduce memory allocations and enhance performance