@pointlessone 2021-01-12
Thank you for your work.

One thing I noticed in the numbers is that prawn-dev uses more memory. I assume prawn-dev represents this PR because timing seem to roughly match what you provide in Run Time section. Is this peak memory, total allocated or something else?