@gettalong 2021-01-12
Yes, prawn-dev is the modified version with the code from this PR and the one from the linked pdf-core PR.

The memory column is the maximum resident memory according to `/usr/bin/time`.

I have run the benchmark again with a similar result, i.e. the memory usage is slightly higher. However, I can't really say why this is the case.