@gettalong 2021-01-17
The pull request https://github.com/prawnpdf/ttfunk/pull/90 makes the TrueType version of the benchmark faster by reducing object allocations by about 69%.