@lucaong 2021-04-06
Interestingly, while debugging why generating a certain large PDF in our app was taking so long, I discovered that it was allocating an unexpectedly high number of objects. We later identified the same issues that this PR and the linked one to `pdf-core` are addressing as the main culprits. I was about to prepare a PR when I noticed this one.

Is there a plan to merge it? If so, is there any way I can help with it?