@aried3r 2021-07-15
We are also facing a certain slowdown to the point that people are starting to notice when generating bigger PDFs.

Would it be possible to rebase this, @gettalong or re-run the CI workflow, @pointlessone? Sadly the CI logs have expired and I cannot see why they failed.