@aried3r 2021-07-15
Thank you both for the quick responses! 

We'll point our `Gemfile` to the main branches of prawn-pdf and pdf-core until new releases are cut.